Backend/requirement.txt
flask
flask-cors
sqlalchemy
psycopg2-binary
python-dotenv
requests
beautifulsoup4
langchain
langchain-google-genai
